sentiment/requirements.txt
selenium
webdriver-manager
pandas
nltk
textblob
transformers
python-dotenv
schedule
